combine economic & governance
